wallet type fix
